input: synaptics_dsx: Handle more framebuffer hints
The driver currently only tries to honor doze mode by watching
for one of the low power mode hints and assuming all other modes
are not doze.  Recent changes to Ambient display have taken
advantage of additional power states, which we wrongly assumed
should not be using the touchpanel.  Fix this by handling all of
the currently supported power states.

* Adapted from Pat Erley <perley@cyngn.com>'s change on bacon:
  https://github.com/LineageOS/android_kernel_oneplus_msm8974/commit/baa84b1ac8893865f8580999bc1ac4118a52e1cb

Change-Id: If976ddbc30761fa3bb973f25ba66a8a7e3fa29dd
Signed-off-by: Paul Keith <javelinanddart@gmail.com>

input: synaptics_dsx: Fix crashes due to fb hints code

When updating this driver to handle new power modes, the data
pointer of the fb_event was incorrectly being dereferenced for
all events.

This data pointer may only be dereferenced for events that are
known to populate it (additionally, you need to know the event
to understand what type of data it points to).

Fix these crashes by only dereferencing the pointer for the
correct events.

Change-Id: I12d67c451b7059dfcbc15c64bd81b7a9684b0cdf